[TTAHUB-5488] Refactor security findings tooling and stack cleanup (#3790)
* Bump fast-xml-builder from 1.1.5 to 1.2.0

Bumps [fast-xml-builder](https://github.com/NaturalIntelligence/fast-xml-builder) from 1.1.5 to 1.2.0.
- [Changelog](https://github.com/NaturalIntelligence/fast-xml-builder/blob/main/CHANGELOG.md)
- [Commits](https://github.com/NaturalIntelligence/fast-xml-builder/compare/v1.1.5...v1.2.0)

---
updated-dependencies:
- dependency-name: fast-xml-builder
  dependency-version: 1.2.0
  dependency-type: indirect
...


* [TTAHUB-5488] Refactor security findings tooling and stack cleanup

* Potential fix for pull request finding


* track pending linkify-it advisory and dedupe validation errors

---------
